Upgrade to proper AST manipuations
Old function parsing code was pretty dodgy and failed in a lot of
places. This makes the parsing code more robust. There's still more work
to be done (such as tracking local variables in scopes etc) but the main
thread ticking is back in action (bar some statements that still need
to be worked on).